chore: Move domain classes in domain package
