+++ src/main/kotlin/blackjack/domain/BlackJackGame.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 import blackjack.view.PlayerResultDto
