+++ src/main/kotlin/blackjack/domain/BlackjackResultMaker.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 import blackjack.view.PlayerResultDto
