+++ src/main/kotlin/blackjack/domain/Point.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 @JvmInline
 value class Point(val value: Int) {
