+++ src/test/kotlin/blackjack/CardDeckTest.kt
@@ -1,5 +1,6 @@
 package blackjack
 
+import blackjack.domain.CardDeck
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.throwable.shouldHaveMessage
