+++ src/test/kotlin/blackjack/CardTest.kt
@@ -1,5 +1,6 @@
 package blackjack
 
+import blackjack.domain.Card
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
