+++ src/test/kotlin/blackjack/PointTest.kt
@@ -1,5 +1,6 @@
 package blackjack
 
+import blackjack.domain.Point
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
